Added authorization to validate creator/user

+++ routes/posts.js
@@ -44,6 +44,10 @@ var getPostModel = req => {
     setImagePath(postModel, req);
   }
 
+  if (req.userData) {
+    postModel.creator = req.userData.userId;
+  }
+
   return postModel;
 };
 
@@ -76,8 +80,12 @@ router.patch("/:id",
   multer({ storage: storage }).single("image"), 
   (req, res, next) => {
     const postModel = getPostModel(req);
-    Post.updateOne({ _id: req.params.id }, postModel).then(result => {
-      res.status(200).json({ message: "ok" });
+    Post.updateOne({ _id: req.params.id, creator: postModel.creator }, postModel).then(result => {
+      if (result.nModified > 0) {
+        res.status(200).json({ message: "ok" });
+      } else {
+        res.status(401).json({ message: "Not authorized" });
+      }
     });
   }
 );
@@ -119,8 +127,12 @@ router.get("/:id", (req, res, next) => {
 });
 
 router.delete("/:id", checkAuth, (req, res, next) => {
-  Post.deleteOne({ _id: req.params.id }).then(result => {
-    res.status(200).json(result);
+  Post.deleteOne({ _id: req.params.id, creator: req.userData.userId }).then(result => {
+    if (result.n > 0) {
+      res.status(200).json(result);
+    } else {
+      res.status(401).json({ message: "Not authorized" });
+    }
   });
 });
